backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
gunicorn==21.2.0
motor==3.3.2
pymongo==4.6.0
pydantic==2.5.0
pydantic-settings==2.1.0
pydantic[email]==2.5.0
python-dotenv==1.0.0
python-multipart==0.0.6
PyJWT==2.11.0
passlib[bcrypt]==1.7.4
twilio==8.11.0
slowapi==0.1.9
pytest==7.4.3
httpx==0.25.2
